CMPT 120 Final Project upload
The 4 source code files for my CMPT 120 Machine Learning Final Project and it's requisite data files. This project was rather challenging to do just because of the strict criteria that was imposed on us when doing it. We weren't allowed to use practically any libraries or extensions, so we had to use their provided prediction library, and basically stock python.