Upload Sales Chatbot Project
a simple python chatbot that was started with the intentions of giving me additional practice in python. the bot runs through the typical questions that one may ask while working sales at COBS (a bakery chain in Canada). the goals for the feature set of this bot are to be able to greet the user, take repeated inputs for the users order as many times as the user would like, ask further specific questions depending on the user's input, tally up the total and calculate tax on the order, ask the user for their preferred payment method, optionally print a receipt, and finally save the order history to a data base or a txt file containing all the previous orders.